+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - FilterModule:
     - A new attribute clearAll for clearing all the filtered results after clicking button "Reset All"
 - api/highlightFeaturesByAttribute URL mechanism to query WFS data by parameters and configuration
+- Add possibility to display an error message for an inactive control.
 
 ### Changed